requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
cryptography==41.0.5
defusedxml==0.8.0rc2
distro==1.9.0
dj-rest-auth==5.0.1
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docx2pdf==0.1.8
drf-yasg==1.21.7
etelemetry==0.3.1
filelock==3.13.1
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
looseversion==1.3.0
lxml==5.1.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
paypalrestsdk==1.1.0
Pillow==10.1.0
plum-dispatch==1.7.4
prov==2.0.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydot==2.0.0
PyJWT==2.8.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
# pywin32==306
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
stripe==7.4.0
tqdm==4.66.2
traits==6.3.2
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
drf-spectacular==0.27.2
# Additional dependencies for profile and OCR functionality
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
python-docx>=0.8.11
# PostgreSQL and vector database dependencies
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy>=1.24.0
dj-database-url==2.1.0
